fix bugs

+++ firebase/firebaseConfig.js
@@ -1,8 +1,8 @@
   // Import the functions you need from the SDKs you need
   import { initializeApp } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-app.js";
   import { getAnalytics } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-analytics.js";
-  import { getFirestore, collection, addDoc,  doc, setDoc, getDocs, deleteDoc } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-firestore.js";
-
+  import { getAuth,createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, deleteUser } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-auth.js";
+  import { getFirestore, collection, addDoc,  doc, setDoc, getDocs, deleteDoc, query, where, } from "https://www.gstatic.com/firebasejs/12.13.0/firebase-firestore.js";
 
   // TODO: Add SDKs for Firebase products that you want to use
   // https://firebase.google.com/docs/web/setup#available-libraries
@@ -23,9 +23,11 @@
   const app = initializeApp(firebaseConfig);
   const analytics = getAnalytics(app);
 
-  
 // Initialize Cloud Firestore and get a reference to the service
 const db = getFirestore(app);
 
+// Initialize Firebase Authentication and get a reference to the service
+const auth = getAuth(app);
 
-export { collection, addDoc, db, getFirestore,  doc, setDoc, getDocs, deleteDoc }
+
+export { collection, addDoc, db, getFirestore,  doc, setDoc, getDocs, deleteDoc, auth, getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword, signOut, deleteUser, query, where }
